PowerShell script for scripting out DacPac files from all the databases in your registered servers that have changed
This takes all the servers in the list of Local Server Groups and
Central Management Server groups in your SSMS.  For every database, it
creates a DacPac in a central archive. Having done that, each time it
runs it compares the dacpac with the live database and creates a new
Dacpac and reports the differences if it detects a difference.